Fixed accessor errors
