+++ src/main/java/me/zeroeightsix/kami/manager/managers/MessageManager.kt
@@ -3,7 +3,7 @@ package me.zeroeightsix.kami.manager.managers
 import me.zeroeightsix.kami.event.events.PacketEvent
 import me.zeroeightsix.kami.event.events.SafeTickEvent
 import me.zeroeightsix.kami.manager.Manager
-import me.zeroeightsix.kami.mixin.client.accessor.network.packetMessage
+import me.zeroeightsix.kami.mixin.extension.packetMessage
 import me.zeroeightsix.kami.module.Module
 import me.zeroeightsix.kami.module.modules.client.ChatSetting
 import me.zeroeightsix.kami.util.TaskState
