--- src/main/java/me/zeroeightsix/kami/mixin/client/accessor/AccessorItemTool.kt
@@ -1,5 +1,0 @@
-package me.zeroeightsix.kami.mixin.client.accessor
-
-import net.minecraft.item.ItemTool
-
-val ItemTool.attackDamage get() = (this as AccessorItemTool).attackDamage
