--- src/main/java/me/zeroeightsix/kami/mixin/client/accessor/network/AccessorSPacketEntityVelocity.kt
@@ -1,21 +1,0 @@
-package me.zeroeightsix.kami.mixin.client.accessor.network
-
-import net.minecraft.network.play.server.SPacketEntityVelocity
-
-var SPacketEntityVelocity.packetMotionX: Int
-    get() = this.motionX
-    set(value) {
-        (this as AccessorSPacketEntityVelocity).setMotionX(value)
-    }
-
-var SPacketEntityVelocity.packetMotionY: Int
-    get() = this.motionY
-    set(value) {
-        (this as AccessorSPacketEntityVelocity).setMotionY(value)
-    }
-
-var SPacketEntityVelocity.packetMotionZ: Int
-    get() = this.motionZ
-    set(value) {
-        (this as AccessorSPacketEntityVelocity).setMotionZ(value)
-    }
