--- src/main/java/me/zeroeightsix/kami/mixin/client/accessor/network/AccessorSPacketPosLook.kt
@@ -1,15 +1,0 @@
-package me.zeroeightsix.kami.mixin.client.accessor.network
-
-import net.minecraft.network.play.server.SPacketPlayerPosLook
-
-var SPacketPlayerPosLook.rotationYaw: Float
-    get() = this.yaw
-    set(value) {
-        (this as AccessorSPacketPosLook).setYaw(value)
-    }
-
-var SPacketPlayerPosLook.rotationPitch: Float
-    get() = this.pitch
-    set(value) {
-        (this as AccessorSPacketPosLook).setPitch(value)
-    }
