+++ src/main/java/me/zeroeightsix/kami/module/modules/misc/ColorSign.kt
@@ -1,8 +1,8 @@
 package me.zeroeightsix.kami.module.modules.misc
 
 import me.zeroeightsix.kami.event.events.GuiScreenEvent
-import me.zeroeightsix.kami.mixin.client.accessor.gui.edtiLine
-import me.zeroeightsix.kami.mixin.client.accessor.gui.tileSign
+import me.zeroeightsix.kami.mixin.extension.editLine
+import me.zeroeightsix.kami.mixin.extension.tileSign
 import me.zeroeightsix.kami.module.Module
 import me.zeroeightsix.kami.util.event.listener
 import net.minecraft.client.gui.GuiButton
@@ -28,7 +28,7 @@ object ColorSign : Module() {
         @Throws(IOException::class)
         override fun actionPerformed(button: GuiButton) {
             if (button.id == 0) {
-                tileSign.signText[edtiLine] = TextComponentString(tileSign.signText[edtiLine].formattedText.replace("(§)(.)".toRegex(), "$1$1$2$2"))
+                tileSign.signText[editLine] = TextComponentString(tileSign.signText[editLine].formattedText.replace("(§)(.)".toRegex(), "$1$1$2$2"))
             }
             super.actionPerformed(button)
         }
@@ -36,9 +36,9 @@ object ColorSign : Module() {
         @Throws(IOException::class)
         override fun keyTyped(typedChar: Char, keyCode: Int) {
             super.keyTyped(typedChar, keyCode)
-            var s = (tileSign.signText[edtiLine] as TextComponentString).text
+            var s = (tileSign.signText[editLine] as TextComponentString).text
             s = s.replace('&', '§')
-            tileSign.signText[edtiLine] = TextComponentString(s)
+            tileSign.signText[editLine] = TextComponentString(s)
         }
     }
 }
