+++ src/main/java/me/zeroeightsix/kami/module/modules/movement/AntiHunger.kt
@@ -1,7 +1,7 @@
 package me.zeroeightsix.kami.module.modules.movement
 
 import me.zeroeightsix.kami.event.events.PacketEvent
-import me.zeroeightsix.kami.mixin.client.accessor.network.onGround
+import me.zeroeightsix.kami.mixin.extension.onGround
 import me.zeroeightsix.kami.module.Module
 import me.zeroeightsix.kami.setting.Settings
 import me.zeroeightsix.kami.util.event.listener
