+++ src/main/java/me/zeroeightsix/kami/module/modules/movement/FastFall.kt
@@ -1,9 +1,9 @@
 package me.zeroeightsix.kami.module.modules.movement
 
 import me.zeroeightsix.kami.event.events.SafeTickEvent
-import me.zeroeightsix.kami.mixin.client.accessor.entity.isInWeb
-import me.zeroeightsix.kami.mixin.client.accessor.tickLength
-import me.zeroeightsix.kami.mixin.client.accessor.timer
+import me.zeroeightsix.kami.mixin.extension.isInWeb
+import me.zeroeightsix.kami.mixin.extension.tickLength
+import me.zeroeightsix.kami.mixin.extension.timer
 import me.zeroeightsix.kami.module.Module
 import me.zeroeightsix.kami.setting.Setting
 import me.zeroeightsix.kami.setting.Settings
