+++ src/main/java/me/zeroeightsix/kami/module/modules/player/AntiForceLook.kt
@@ -1,8 +1,8 @@
 package me.zeroeightsix.kami.module.modules.player
 
 import me.zeroeightsix.kami.event.events.PacketEvent
-import me.zeroeightsix.kami.mixin.client.accessor.network.rotationPitch
-import me.zeroeightsix.kami.mixin.client.accessor.network.rotationYaw
+import me.zeroeightsix.kami.mixin.extension.rotationPitch
+import me.zeroeightsix.kami.mixin.extension.rotationYaw
 import me.zeroeightsix.kami.module.Module
 import me.zeroeightsix.kami.util.event.listener
 import net.minecraft.network.play.server.SPacketPlayerPosLook
