+++ src/main/java/me/zeroeightsix/kami/module/modules/player/Blink.kt
@@ -3,9 +3,9 @@ package me.zeroeightsix.kami.module.modules.player
 import me.zeroeightsix.kami.event.events.ConnectionEvent
 import me.zeroeightsix.kami.event.events.PacketEvent
 import me.zeroeightsix.kami.event.events.SafeTickEvent
-import me.zeroeightsix.kami.mixin.client.accessor.network.x
-import me.zeroeightsix.kami.mixin.client.accessor.network.y
-import me.zeroeightsix.kami.mixin.client.accessor.network.z
+import me.zeroeightsix.kami.mixin.extension.x
+import me.zeroeightsix.kami.mixin.extension.y
+import me.zeroeightsix.kami.mixin.extension.z
 import me.zeroeightsix.kami.module.Module
 import me.zeroeightsix.kami.setting.Settings
 import me.zeroeightsix.kami.util.event.listener
