+++ src/main/java/me/zeroeightsix/kami/module/modules/player/NoFall.kt
@@ -3,8 +3,8 @@ package me.zeroeightsix.kami.module.modules.player
 import me.zeroeightsix.kami.command.commands.TeleportCommand
 import me.zeroeightsix.kami.event.events.PacketEvent
 import me.zeroeightsix.kami.event.events.SafeTickEvent
-import me.zeroeightsix.kami.mixin.client.accessor.network.onGround
-import me.zeroeightsix.kami.mixin.client.accessor.rightClickMouse
+import me.zeroeightsix.kami.mixin.extension.onGround
+import me.zeroeightsix.kami.mixin.extension.rightClickMouse
 import me.zeroeightsix.kami.module.Module
 import me.zeroeightsix.kami.setting.Settings
 import me.zeroeightsix.kami.util.BlockUtils
