+++ src/main/java/me/zeroeightsix/kami/util/InventoryUtils.kt
@@ -1,6 +1,6 @@
 package me.zeroeightsix.kami.util
 
-import me.zeroeightsix.kami.mixin.client.accessor.player.syncCurrentPlayItem
+import me.zeroeightsix.kami.mixin.extension.syncCurrentPlayItem
 import net.minecraft.client.Minecraft
 import net.minecraft.inventory.ClickType
 import net.minecraft.item.Item
